[client] ds: avoid util_cursorToInt when warping pointer
Using util_cursorToInt messes with the error tracking for normal movements,
and is not necessary since we are computing an absolute position on the
client window.

Instead, we should pass doubles directly to display servers and let them
decide how to best handle them. For example, XIWarpPointer accepts doubles
directly.

+++ client/displayservers/Wayland/wayland.h
@@ -230,7 +230,7 @@ void waylandCapturePointer(void);
 void waylandUncapturePointer(void);
 void waylandRealignPointer(void);
 void waylandWarpPointer(int x, int y, bool exiting);
-void waylandGuestPointerUpdated(double x, double y, int localX, int localY);
+void waylandGuestPointerUpdated(double x, double y, double localX, double localY);
 
 // output module
 bool waylandOutputInit(void);
